refactor: move API views into temples/api/views.py and clean up temples/views.py
